+++ src/arpes/correction/fermi_edge.py
@@ -9,9 +9,9 @@ import numpy as np
 import xarray as xr
 from matplotlib.axes import Axes
 
+from arpes.correction.intensity_map import shift_by
 from arpes.fits import GStepBModel, LinearModel, QuadraticModel, broadcast_model
 from arpes.provenance import Provenance, provenance, update_provenance
-from arpes.utilities.math import shift_by
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Hashable, Iterable