+++ tests/test_xarray_extensions.py
@@ -300,29 +300,72 @@ class TestGeneralforDataArray:
             "phi": -0.2910254835234106,
         }
 
-    def test_G_shift(
-        self,
-        dataarray_map: xr.DataArray,
-    ) -> None:
-        """Test for G.shift_by."""
+    @pytest.fixture
+    def edge(self, dataarray_map: xr.DataArray) -> xr.DataArray:
         fmap = dataarray_map
         cut = fmap.sum("theta", keep_attrs=True).sel(eV=slice(-0.2, 0.1), phi=slice(-0.25, 0.3))
         fit_results = broadcast_model(AffineBroadenedFD, cut, "phi")
-        edge = (
+        return (
             QuadraticModel()
             .guess_fit(
                 fit_results.results.F.p("center"),
             )
             .eval(x=fmap.phi)
         )
+
+    def test_G_shift(
+        self,
+        dataarray_map: xr.DataArray,
+        edge: xr.DataArray,
+    ) -> None:
+        """Test for G.shift_by."""
+        fmap = dataarray_map
         np.testing.assert_allclose(
-            actual=fmap.G.shift_by(edge, shift_axis="eV", by_axis="phi").sel(
+            actual=fmap.G.shift_by(
+                edge,
+                shift_axis="eV",
+                by_axis="phi",
+            ).sel(
                 eV=0,
                 method="nearest",
             )[:][0][:5],
             desired=np.array([5.625749, 566.8711542, 757.8334417, 637.2900199, 610.679927]),
             rtol=1e-2,
         )
+
+    def test_G_shift_with_extend_coords(
+        self,
+        dataarray_map: xr.DataArray,
+        edge: xr.DataArray,
+    ) -> None:
+        """Test for G.shift_by with extend_coords."""
+        fmap = dataarray_map
+        shifted_without_extension_coords = fmap.G.shift_by(
+            edge,
+            shift_axis="eV",
+            by_axis="phi",
+            extend_coords=False,
+        )
+        assert shifted_without_extension_coords.shape == (81, 150, 111)
+
+        shifted_with_extension_coords = fmap.G.shift_by(
+            edge,
+            shift_axis="eV",
+            by_axis="phi",
+            extend_coords=True,
+        )
+
+        assert shifted_with_extension_coords.shape == (81, 154, 111)
+
+        should_be_zero_array = (
+            (
+                shifted_with_extension_coords.isel(eV=slice(0, 150))
+                - shifted_without_extension_coords
+            )
+            .fillna(0)
+            .values
+        )
+        assert np.all(should_be_zero_array == 0)
 
     def test_G_meshgrid(self, dataarray_cut: xr.DataArray) -> None:
         """Test for G.meshgrid, G.scale_meshgrid, G.shift_meshgrid."""